Conflicts:
	BLL/Account/AccountBLL.cs
	BLL/Category/CategoryBLL.cs
	BLL/Image/IImageLogic.cs
	BLL/Order/IOrderLogic.cs
	BusinessLogic/Account/IAccountLogic.cs
	BusinessLogic/App.config
	BusinessLogic/BusinessLogic.csproj
	BusinessLogic/Category/CategoryLogic.cs
	BusinessLogic/Image/ImageBLL.cs
	BusinessLogic/Order/OrderBLL.cs
	BusinessLogic/Product/IProductLogic.cs
	BusinessLogic/Product/ProductBLL.cs
	DAL/Account/AccountRepo.cs
	DAL/Account/IAccountRepo.cs
	DAL/App.config
	DAL/Category/CategoryRepo.cs
	DAL/Category/CategoryRepoStub.cs
	DAL/Category/ICategoryRepo.cs
	DAL/Image/IImageRepo.cs
	DAL/Image/ImageRepoStub.cs
	DAL/Order/IOrderRepo.cs
	DAL/Order/OrderRepo.cs
	DAL/TankshopDbContext.cs
	DataAccess/Account/AccountRepositoryStub.cs
	DataAccess/Image/ImageRepo.cs
	DataAccess/Migrations/Configuration.cs
	DataAccess/Product/IProductRepo.cs
	DataAccess/Product/ProductRepo.cs
	DataAccess/Product/ProductRepoStub.cs
	DataAccess/packages.config
	Logging/LogHandler.cs
	Logging/Logging.csproj
	Logging/Properties/AssemblyInfo.cs
	Model/AdminModel.cs
	Model/CategoryModel.cs
	Model/CustomerModel.cs
	Model/ImageModel.cs
	Model/Model.csproj
	Model/OrderModel.cs
	Model/Orderline.cs
	Model/PersonModel.cs
	Model/ProductModel.cs
	Nettbutikk.sln
	Nettbutikk/Content/Site.css
	Nettbutikk/Controllers/AccountController.cs
	Nettbutikk/Controllers/CartController.cs
	Nettbutikk/Controllers/CategoryController.cs
	Nettbutikk/Controllers/CheckoutController.cs
	Nettbutikk/CookieHandler.cs
	Nettbutikk/Global.asax.cs
	Nettbutikk/Models/ProductView.cs
	Nettbutikk/Oblig1_Nettbutikk.csproj
	Nettbutikk/Properties/AssemblyInfo.cs
	Nettbutikk/Views/Account/_CustomerEditInfoPartial.cshtml
	Nettbutikk/Views/Account/_CustomerListOrdersPartial.cshtml
	Nettbutikk/Views/Category/CreateCategory.cshtml
	Nettbutikk/Views/Home/Index.cshtml
	Nettbutikk/Views/Product/Product_Details.cshtml
	Nettbutikk/Views/Shared/_HeaderPartial.cshtml
	Nettbutikk/Views/Shared/_LoginModal.cshtml
	Nettbutikk/Views/Shared/_RegisterModal.cshtml
	Nettbutikk/Views/Web.config
	Nettbutikk/Web.config
	Oblig1_Nettbutikk/Controllers/HomeController.cs
	Oblig1_Nettbutikk/Controllers/ImageController.cs
	Oblig1_Nettbutikk/Controllers/ProductController.cs
	Oblig1_Nettbutikk/Logs/log.txt
	Oblig1_Nettbutikk/Migrations/Configuration.cs
	Oblig1_Nettbutikk/Nettbutikk.csproj
	Oblig1_Nettbutikk/Oblig1_Nettbutikk.csproj
	Oblig1_Nettbutikk/Views/Home/Product_ListItem.cshtml
	Oblig1_Nettbutikk/Views/Product/Product.cshtml
	TankShopTest/Controllers/ImageControllerTest.cs
	TankShopTest/Controllers/ProductControllerTest.cs
	TankShopTest/TankShopTest.csproj
	TankShopTest/packages.config
	TankShopUnitTest/app.config
	readme.txt

+++ BusinessLogic/Category/CategoryBLL.cs
@@ -4,13 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Nettbutikk.Model;
-using Nettbutikk.DAL;
+using DAL.Category;
 
-namespace Nettbutikk.BLL
+namespace BLL.Category
 {
     public class CategoryBLL :  ICategoryLogic
     {
-
+        
         private ICategoryRepo repo;
 
         public CategoryBLL()
@@ -23,49 +23,35 @@ namespace Nettbutikk.BLL
             this.repo = repo;
         }
 
-        //public List<CategoryModel> GetAllCategoryModels()
-        //{
-        //    return repo.GetAllCategoryModels();
-        //}
 
         public bool AddCategory(string Name)
         {
             return repo.AddCategory(Name);
         }
 
-        public bool DeleteCategory(int CategoryId)
+        public bool DeleteCategory(int CategoryId, int AdminId)
         {
-            //Category category = repo.GetCategory(CategoryId);
+            Nettbutikk.Model.Category category = repo.GetCategory(CategoryId);
 
-            //if (category == null)
-            //    return false;
+            if (category == null)
+                return false;
 
-            //if (!repo.AddOldCategory(category.Name, 1))//Get admin id from session
-            //    return false;
+            if (!repo.AddOldCategory(category.Name, AdminId))
+                return false;
 
             return repo.DeleteCategory(CategoryId);
         }
 
-        //public List<Category> GetAllCategories()
-        //{
-        //    return repo.GetAllCategories();
-        //}
-
-        //public Category GetCategory(int CategoryId)
-        //{
-        //    return repo.GetCategory(CategoryId);
-        //}
-
-        public bool UpdateCategory(int CategoryId, string Name)
+        public bool UpdateCategory(int CategoryId, string Name, int AdminId)
         {
 
-            //Category category = repo.GetCategory(CategoryId);
+            Nettbutikk.Model.Category category = repo.GetCategory(CategoryId);
 
-            //if (category == null)
-            //    return false;
+            if (category == null)
+                return false;
 
-            //if (!repo.AddOldCategory(category.Name, 1))//Get admin id from session
-            //    return false;
+            if (!repo.AddOldCategory(category.Name, AdminId))
+                return false;
 
             return repo.UpdateCategory(CategoryId, Name);
         }
@@ -75,19 +61,32 @@ namespace Nettbutikk.BLL
             return repo.GetCategoryName(CategoryId);
         }
 
-        public List<CategoryModel> GetAllCategories()
-        {
-            return repo.GetAllCategories();
-        }
-
-        public CategoryModel GetCategory(int CategoryId)
-        {
-            return repo.GetCategory(CategoryId);
-        }
 
         public int FirstCategoryWithProducts()
         {
             return repo.FirstCategoryWithProducts();
         }
+
+        public Nettbutikk.Model.Category GetCategory(int CategoryId)
+        {
+            return repo.GetCategory(CategoryId);
+        }
+
+        public List<Nettbutikk.Model.Category> GetAllCategories()
+        {
+            return repo.GetAllCategories();
+        }
+
+
+        public CategoryModel GetCategoryModel(int CategoryId)
+        {
+            return repo.GetCategoryModel(CategoryId);
+        }
+
+        public List<CategoryModel> GetAllCategoryModels()
+        {
+            return repo.GetAllCategoryModels();
+        }
+
     }
 }
